Move utility methods out of KIDL that aren't used there
Post parse utility functions belong in SDK code, not in KIDL code. The
moved functions are not called anywhere in KIDL code
