+++ src/java/us/kbase/mobu/compiler/JavaTypeGenerator.java
@@ -40,7 +40,6 @@ import us.kbase.kidl.KbTypedef;
 import us.kbase.kidl.KbUnspecifiedObject;
 import us.kbase.kidl.KidlParseException;
 import us.kbase.kidl.KidlParser;
-import us.kbase.kidl.Utils;
 import us.kbase.mobu.util.DiskFileSaver;
 import us.kbase.mobu.util.FileSaveCodeWriter;
 import us.kbase.mobu.util.FileSaver;
